Split up trakt service into smaller service classes.
