+++ android-core/workmanager/src/main/kotlin/com/thomaskioko/tvmaniac/workmanager/SyncDiscoverShowsWorker.kt
@@ -3,7 +3,7 @@ package com.thomaskioko.tvmaniac.workmanager
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.thomaskioko.tvmaniac.shows.api.ShowsRepository
+import com.thomaskioko.tvmaniac.shows.api.DiscoverRepository
 import com.thomaskioko.tvmaniac.util.KermitLogger
 import me.tatarka.inject.annotations.Assisted
 import me.tatarka.inject.annotations.Inject
@@ -13,7 +13,7 @@ class SyncDiscoverShowsWorker(
     @Assisted context: Context,
     @Assisted workerParameters: WorkerParameters,
     private val logger: KermitLogger,
-    private val showsRepository: ShowsRepository,
+    private val discoverRepository: DiscoverRepository,
 ) : CoroutineWorker(context, workerParameters) {
 
     companion object {
@@ -22,7 +22,7 @@ class SyncDiscoverShowsWorker(
 
     override suspend fun doWork(): Result {
         logger.debug("$tags worker running")
-        showsRepository.fetchDiscoverShows()
+        discoverRepository.fetchDiscoverShows()
         return Result.success()
     }
 }
