+++ core/trakt-api/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/api/TraktTokenRemoteDataSource.kt
@@ -1,0 +1,13 @@
+package com.thomaskioko.tvmaniac.trakt.api
+
+import com.thomaskioko.tvmaniac.trakt.api.model.TraktAccessRefreshTokenResponse
+import com.thomaskioko.tvmaniac.trakt.api.model.TraktAccessTokenResponse
+
+interface TraktTokenRemoteDataSource {
+
+    suspend fun getAccessToken(authCode: String): TraktAccessTokenResponse
+
+    suspend fun getAccessRefreshToken(refreshToken: String): TraktAccessRefreshTokenResponse
+
+    suspend fun revokeAccessToken(authCode: String)
+}
