+++ core/trakt-api/implementation/src/androidMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktPlatformComponent.kt
@@ -1,5 +1,6 @@
-package com.thomaskioko.trakt.service.implementation
+package com.thomaskioko.trakt.service.implementation.inject
 
+import com.thomaskioko.trakt.service.implementation.TraktAuthInterceptor
 import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
 import io.ktor.client.engine.okhttp.OkHttp
 import me.tatarka.inject.annotations.Provides
