+++ core/trakt-api/implementation/src/commonMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktPlatformComponent.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.trakt.service.implementation
+package com.thomaskioko.trakt.service.implementation.inject
 
 import io.ktor.client.engine.HttpClientEngine
