+++ data/seasondetails/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasondetails/implementation/SeasonDetailsRepositoryImpl.kt
@@ -11,7 +11,7 @@ import com.thomaskioko.tvmaniac.core.networkutil.networkBoundResult
 import com.thomaskioko.tvmaniac.episodes.api.EpisodesDao
 import com.thomaskioko.tvmaniac.seasondetails.api.SeasonDetailsDao
 import com.thomaskioko.tvmaniac.seasondetails.api.SeasonDetailsRepository
-import com.thomaskioko.tvmaniac.trakt.api.TraktRemoteDataSource
+import com.thomaskioko.tvmaniac.trakt.api.TraktShowsRemoteDataSource
 import com.thomaskioko.tvmaniac.trakt.api.model.ErrorResponse
 import com.thomaskioko.tvmaniac.trakt.api.model.TraktSeasonEpisodesResponse
 import com.thomaskioko.tvmaniac.util.KermitLogger
@@ -23,7 +23,7 @@ import me.tatarka.inject.annotations.Inject
 
 @Inject
 class SeasonDetailsRepositoryImpl(
-    private val traktRemoteDataSource: TraktRemoteDataSource,
+    private val remoteDataSource: TraktShowsRemoteDataSource,
     private val seasonCache: SeasonDetailsDao,
     private val episodesDao: EpisodesDao,
     private val exceptionHandler: NetworkExceptionHandler,
@@ -35,7 +35,7 @@ class SeasonDetailsRepositoryImpl(
         networkBoundResult(
             query = { seasonCache.observeShowEpisodes(traktId) },
             shouldFetch = { it.isNullOrEmpty() },
-            fetch = { traktRemoteDataSource.getSeasonEpisodes(traktId) },
+            fetch = { remoteDataSource.getSeasonEpisodes(traktId) },
             saveFetchResult = { mapResponse(traktId, it) },
             exceptionHandler = exceptionHandler,
             coroutineDispatcher = dispatcher.io,
