+++ data/seasons/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/implementation/SeasonsStore.kt
@@ -5,7 +5,7 @@ import com.thomaskioko.tvmaniac.core.networkutil.ApiResponse
 import com.thomaskioko.tvmaniac.resourcemanager.api.LastRequest
 import com.thomaskioko.tvmaniac.resourcemanager.api.RequestManagerRepository
 import com.thomaskioko.tvmaniac.seasons.api.SeasonsDao
-import com.thomaskioko.tvmaniac.trakt.api.TraktRemoteDataSource
+import com.thomaskioko.tvmaniac.trakt.api.TraktShowsRemoteDataSource
 import com.thomaskioko.tvmaniac.util.KermitLogger
 import com.thomaskioko.tvmaniac.util.model.AppCoroutineScope
 import kotlinx.datetime.Clock
@@ -17,14 +17,14 @@ import org.mobilenativefoundation.store.store5.StoreBuilder
 
 @Inject
 class SeasonsStore(
-    private val traktRemoteDataSource: TraktRemoteDataSource,
+    private val remoteDataSource: TraktShowsRemoteDataSource,
     private val requestManagerRepository: RequestManagerRepository,
     private val seasonsDao: SeasonsDao,
     private val scope: AppCoroutineScope,
     private val logger: KermitLogger,
 ) : Store<Long, List<Seasons>> by StoreBuilder.from<Long, List<Seasons>, List<Seasons>>(
     fetcher = Fetcher.of { id ->
-        when (val response = traktRemoteDataSource.getShowSeasons(id)) {
+        when (val response = remoteDataSource.getShowSeasons(id)) {
             is ApiResponse.Success -> response.body.toSeasonCacheList(id)
             is ApiResponse.Error.GenericError -> {
                 logger.error("SeasonsStore GenericError", "$response")
