+++ data/similar/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/similar/implementation/SimilarShowStore.kt
@@ -6,7 +6,7 @@ import com.thomaskioko.tvmaniac.resourcemanager.api.LastRequest
 import com.thomaskioko.tvmaniac.resourcemanager.api.RequestManagerRepository
 import com.thomaskioko.tvmaniac.shows.api.ShowsDao
 import com.thomaskioko.tvmaniac.similar.api.SimilarShowsDao
-import com.thomaskioko.tvmaniac.trakt.api.TraktRemoteDataSource
+import com.thomaskioko.tvmaniac.trakt.api.TraktShowsRemoteDataSource
 import com.thomaskioko.tvmaniac.util.KermitLogger
 import com.thomaskioko.tvmaniac.util.model.AppCoroutineScope
 import me.tatarka.inject.annotations.Inject
@@ -17,7 +17,7 @@ import org.mobilenativefoundation.store.store5.StoreBuilder
 
 @Inject
 class SimilarShowStore(
-    private val traktRemoteDataSource: TraktRemoteDataSource,
+    private val remoteDataSource: TraktShowsRemoteDataSource,
     private val similarShowsDao: SimilarShowsDao,
     private val showsDao: ShowsDao,
     private val requestManagerRepository: RequestManagerRepository,
@@ -26,7 +26,7 @@ class SimilarShowStore(
 ) : Store<Long, List<SimilarShows>> by StoreBuilder.from<Long, List<SimilarShows>, List<SimilarShows>>(
     fetcher = Fetcher.of { id ->
 
-        when (val apiResult = traktRemoteDataSource.getSimilarShows(id)) {
+        when (val apiResult = remoteDataSource.getSimilarShows(id)) {
             is ApiResponse.Success -> apiResult.body.responseToShow()
 
             is ApiResponse.Error.GenericError -> {
